refactor: decompose ExtractedIssue.swift into 5 domain-scoped files (#641-#645)
Splits the 772-line Models/ExtractedIssue.swift monolith (which contained 20+ types
across 5 domains) into 5 domain-scoped files + a renamed 6-line utilities remnant.

## Slices bundled in this PR

Each slice was originally filed as a separate ticket (#641-#645). Bundled into one
PR because all 5 touch the same source file — sequential in-file extraction is
cleaner than 5 parallel worktrees rebasing against each other.

**Slice A (#641)** — IssueCore.swift (351 lines):
- ExtractedIssueCategory, ExtractedIssueSeverity (enums)
- IssueReproductionStep, IssueScreenshotAnnotation, ExtractedIssue (structs)
- SHA-256: f6344158f35d3289a596e1b11a681aaccac9aa50a023585e5528c37fcfb311fa

**Slice B (#642)** — IssueExportTargets.swift (122 lines):
- GitHubIssueExportTarget, JiraIssueExportTarget, IssueExtractionResult
- ExportDestination, ExportResult
- SHA-256: ecb79d344852a4f26244bf6a81e19a9d2d1fe666496aba1e2a2ce95243aa4b1f

**Slice C (#643)** — IssueExportReview.swift (128 lines):
- SimilarIssueResolution, SimilarIssueMatch
- IssueExportReviewItem, IssueExportReview
- SHA-256: 807f66a7cbcabaf312a2fc1aeee16a218c46b4d01a8cdd1a3a3e68d3f81f9279

**Slice D (#644)** — GitHubExportConfig.swift (72 lines):
- GitHubExportConfiguration, GitHubRepositoryOption
- SHA-256: 382deccb089b11d28f92a02b1853d81157c2458c262c5362a5fc4eb1e64a2c9d

**Slice E (#645)** — JiraExportConfig.swift (90 lines):
- JiraConnectionConfiguration, JiraExportConfiguration
- JiraProjectOption, JiraIssueTypeOption
- SHA-256: c4ae361b46f0d1d0440b503193f95825db7137b5267c852902b5ab2c13ca55f9

## Utilities remnant

After the 5 extractions, ExtractedIssue.swift retained only a 6-line String
extension (nilIfEmpty). Renamed via git mv to StringExtensions.swift for
honest naming — the old name no longer described its contents.

## Test

BugNarratorTests: 667 (unchanged). All byte-identical moves per file's SHA;
git detected the rename to StringExtensions.swift at 100% similarity for the
retained String extension.

Imports: Foundation on each new file (for UUID, Date, Codable).
