+++ Sources/BugNarrator/Models/IssueCore.swift
@@ -351,422 +351,3 @@ struct ExtractedIssue: Identifiable, Codable, Equatable {
     }
 }
 
-struct GitHubIssueExportTarget: Codable, Equatable {
-    var repositoryID: String?
-    var owner: String
-    var repository: String
-    var labels: [String]
-
-    init(
-        repositoryID: String? = nil,
-        owner: String = "",
-        repository: String = "",
-        labels: [String] = []
-    ) {
-        self.repositoryID = repositoryID?.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty
-        self.owner = owner.trimmingCharacters(in: .whitespacesAndNewlines)
-        self.repository = repository.trimmingCharacters(in: .whitespacesAndNewlines)
-        self.labels = labels
-            .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-            .filter { !$0.isEmpty }
-    }
-
-    var isComplete: Bool {
-        !owner.isEmpty && !repository.isEmpty
-    }
-
-    var displayLabel: String {
-        isComplete ? "\(owner)/\(repository)" : "Choose a GitHub repository"
-    }
-}
-
-struct JiraIssueExportTarget: Codable, Equatable {
-    var projectID: String?
-    var projectKey: String
-    var projectName: String?
-    var issueTypeID: String
-    var issueTypeName: String
-
-    init(
-        projectID: String? = nil,
-        projectKey: String = "",
-        projectName: String? = nil,
-        issueTypeID: String = "",
-        issueTypeName: String = ""
-    ) {
-        self.projectID = projectID?.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty
-        self.projectKey = projectKey.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
-        self.projectName = projectName?.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty
-        self.issueTypeID = issueTypeID.trimmingCharacters(in: .whitespacesAndNewlines)
-        self.issueTypeName = issueTypeName.trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-
-    var isComplete: Bool {
-        !projectKey.isEmpty && (!issueTypeID.isEmpty || !issueTypeName.isEmpty)
-    }
-
-    var projectDisplayLabel: String {
-        guard let projectName else {
-            return projectKey.isEmpty ? "Choose a Jira project" : projectKey
-        }
-
-        return "\(projectKey) - \(projectName)"
-    }
-}
-
-struct IssueExtractionResult: Codable, Equatable {
-    let generatedAt: Date
-    var summary: String
-    var guidanceNote: String
-    var issues: [ExtractedIssue]
-
-    init(
-        generatedAt: Date = Date(),
-        summary: String,
-        guidanceNote: String = "Extracted issues are draft suggestions and should be reviewed before export.",
-        issues: [ExtractedIssue]
-    ) {
-        self.generatedAt = generatedAt
-        self.summary = summary
-        self.guidanceNote = guidanceNote
-        self.issues = issues
-    }
-
-    var selectedIssues: [ExtractedIssue] {
-        issues.filter(\.isSelectedForExport)
-    }
-}
-
-enum ExportDestination: String, Codable, CaseIterable, Identifiable {
-    case github = "GitHub"
-    case jira = "Jira"
-
-    var id: String { rawValue }
-
-    var actionTitle: String {
-        "Export to \(rawValue)"
-    }
-}
-
-struct ExportResult: Identifiable, Equatable {
-    let id: UUID
-    let sourceIssueID: UUID
-    let destination: ExportDestination
-    let remoteIdentifier: String
-    let remoteURL: URL?
-    let exportedAt: Date
-
-    init(
-        id: UUID = UUID(),
-        sourceIssueID: UUID,
-        destination: ExportDestination,
-        remoteIdentifier: String,
-        remoteURL: URL?,
-        exportedAt: Date = Date()
-    ) {
-        self.id = id
-        self.sourceIssueID = sourceIssueID
-        self.destination = destination
-        self.remoteIdentifier = remoteIdentifier
-        self.remoteURL = remoteURL
-        self.exportedAt = exportedAt
-    }
-}
-
-enum SimilarIssueResolution: String, CaseIterable, Identifiable {
-    case exportNew = "Export as New"
-    case linkAsRelated = "Link as Related"
-    case markDuplicate = "Mark as Duplicate"
-
-    var id: String { rawValue }
-
-    var helpText: String {
-        switch self {
-        case .exportNew:
-            return "Create a brand-new tracker issue."
-        case .linkAsRelated:
-            return "Create a new issue and reference an existing tracker issue as related context."
-        case .markDuplicate:
-            return "Skip creating a new issue and use the matched tracker issue instead."
-        }
-    }
-}
-
-struct SimilarIssueMatch: Identifiable, Equatable {
-    let id: String
-    let remoteIdentifier: String
-    let title: String
-    let summary: String
-    let remoteURL: URL?
-    let confidence: Double
-    let reasoning: String
-
-    init(
-        remoteIdentifier: String,
-        title: String,
-        summary: String,
-        remoteURL: URL?,
-        confidence: Double,
-        reasoning: String
-    ) {
-        self.id = remoteIdentifier.lowercased()
-        self.remoteIdentifier = remoteIdentifier
-        self.title = title
-        self.summary = summary
-        self.remoteURL = remoteURL
-        self.confidence = min(max(confidence, 0), 1)
-        self.reasoning = reasoning.trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-
-    var confidenceLabel: String {
-        "\(Int((confidence * 100).rounded()))%"
-    }
-}
-
-struct IssueExportReviewItem: Identifiable, Equatable {
-    let id: UUID
-    var issue: ExtractedIssue
-    var matches: [SimilarIssueMatch]
-    var resolution: SimilarIssueResolution
-    var selectedMatchID: String?
-
-    init(
-        issue: ExtractedIssue,
-        matches: [SimilarIssueMatch],
-        resolution: SimilarIssueResolution = .exportNew,
-        selectedMatchID: String? = nil
-    ) {
-        self.id = issue.id
-        self.issue = issue
-        self.matches = matches
-        self.resolution = resolution
-        self.selectedMatchID = selectedMatchID ?? matches.first?.id
-
-        if resolution != .exportNew, self.selectedMatchID == nil {
-            self.selectedMatchID = matches.first?.id
-        }
-    }
-
-    var selectedMatch: SimilarIssueMatch? {
-        guard let selectedMatchID else {
-            return nil
-        }
-
-        return matches.first { $0.id == selectedMatchID }
-    }
-
-    var hasMatches: Bool {
-        !matches.isEmpty
-    }
-
-    mutating func setResolution(_ resolution: SimilarIssueResolution) {
-        self.resolution = resolution
-
-        guard resolution != .exportNew else {
-            return
-        }
-
-        if selectedMatch == nil {
-            selectedMatchID = matches.first?.id
-        }
-    }
-
-    mutating func selectMatch(id: String) {
-        guard matches.contains(where: { $0.id == id }) else {
-            return
-        }
-
-        selectedMatchID = id
-    }
-}
-
-struct IssueExportReview: Identifiable, Equatable {
-    let id: UUID
-    let destination: ExportDestination
-    let sessionID: UUID
-    var items: [IssueExportReviewItem]
-
-    init(
-        destination: ExportDestination,
-        sessionID: UUID,
-        items: [IssueExportReviewItem]
-    ) {
-        self.id = UUID()
-        self.destination = destination
-        self.sessionID = sessionID
-        self.items = items
-    }
-
-    var hasMatches: Bool {
-        items.contains(where: \.hasMatches)
-    }
-}
-
-struct GitHubExportConfiguration: Equatable {
-    let token: String
-    let repositoryID: String?
-    let owner: String
-    let repository: String
-    let labels: [String]
-
-    init(
-        token: String,
-        repositoryID: String? = nil,
-        owner: String,
-        repository: String,
-        labels: [String]
-    ) {
-        self.token = token
-        self.repositoryID = repositoryID
-        self.owner = owner
-        self.repository = repository
-        self.labels = labels
-    }
-
-    var isComplete: Bool {
-        !token.isEmpty && !owner.isEmpty && !repository.isEmpty
-    }
-
-    var targetIdentity: String {
-        repositoryID?.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty
-            ?? "\(owner.lowercased())/\(repository.lowercased())"
-    }
-
-    var issueExportTarget: GitHubIssueExportTarget {
-        GitHubIssueExportTarget(
-            repositoryID: repositoryID,
-            owner: owner,
-            repository: repository,
-            labels: labels
-        )
-    }
-}
-
-struct GitHubRepositoryOption: Equatable, Identifiable {
-    let repositoryID: String
-    let owner: String
-    let name: String
-    let description: String?
-
-    init(
-        repositoryID: String? = nil,
-        owner: String,
-        name: String,
-        description: String?
-    ) {
-        self.repositoryID = repositoryID ?? "\(owner.lowercased())/\(name.lowercased())"
-        self.owner = owner
-        self.name = name
-        self.description = description
-    }
-
-    var id: String { repositoryID }
-
-    var fullName: String {
-        "\(owner)/\(name)"
-    }
-
-    var displayLabel: String {
-        guard let description, !description.isEmpty else {
-            return fullName
-        }
-
-        return "\(fullName) - \(description)"
-    }
-}
-
-struct JiraConnectionConfiguration: Equatable {
-    let baseURL: URL
-    let email: String
-    let apiToken: String
-
-    var isComplete: Bool {
-        !email.isEmpty && !apiToken.isEmpty
-    }
-}
-
-struct JiraExportConfiguration: Equatable {
-    let baseURL: URL
-    let email: String
-    let apiToken: String
-    let projectID: String?
-    let projectKey: String
-    let issueTypeID: String
-    let issueTypeName: String
-
-    init(
-        baseURL: URL,
-        email: String,
-        apiToken: String,
-        projectID: String? = nil,
-        projectKey: String,
-        issueTypeID: String = "",
-        issueTypeName: String? = nil,
-        issueType: String? = nil
-    ) {
-        self.baseURL = baseURL
-        self.email = email
-        self.apiToken = apiToken
-        self.projectID = projectID
-        self.projectKey = projectKey
-        self.issueTypeID = issueTypeID
-        self.issueTypeName = issueTypeName ?? issueType ?? ""
-    }
-
-    var isComplete: Bool {
-        !email.isEmpty
-            && !apiToken.isEmpty
-            && !projectKey.isEmpty
-            && !(issueTypeID.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-                 && issueTypeName.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty)
-    }
-
-    var targetIdentity: String {
-        let normalizedProjectID = projectID?.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty ?? projectKey.uppercased()
-        let normalizedIssueTypeID = issueTypeID.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty
-        let normalizedIssueTypeName = issueTypeName
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-            .lowercased()
-            .nilIfEmpty
-            ?? "unknown"
-        return "\(normalizedProjectID)::\(normalizedIssueTypeID ?? normalizedIssueTypeName)"
-    }
-
-    var issueExportTarget: JiraIssueExportTarget {
-        JiraIssueExportTarget(
-            projectID: projectID,
-            projectKey: projectKey,
-            issueTypeID: issueTypeID,
-            issueTypeName: issueTypeName
-        )
-    }
-}
-
-struct JiraProjectOption: Equatable, Identifiable {
-    let projectID: String
-    let key: String
-    let name: String
-
-    init(projectID: String? = nil, key: String, name: String) {
-        self.projectID = projectID ?? key
-        self.key = key
-        self.name = name
-    }
-
-    var id: String { projectID }
-
-    var displayLabel: String {
-        "\(key) - \(name)"
-    }
-}
-
-struct JiraIssueTypeOption: Equatable, Identifiable {
-    let id: String
-    let name: String
-}
-
-extension String {
-    var nilIfEmpty: String? {
-        isEmpty ? nil : self
-    }
-}
