+++ Sources/BugNarrator/Models/StringExtensions.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension String {
+    var nilIfEmpty: String? {
+        isEmpty ? nil : self
+    }
+}
